Use createSlice from RTK for guitars
