+++ src/helpers/get-mock-store.ts
@@ -1,6 +1,6 @@
 import { configureMockStore } from '@jedmao/redux-mock-store';
 import { RootState } from '../store/store';
-import { initialState as guitarsInitialState } from '../store/guitars/reducer';
+import { initialState as guitarsInitialState } from '../store/guitars/slice';
 import { initialState as paginationInitialState } from '../store/pagination/slice';
 import { initialState as filtersInitialState } from '../store/filters/slice';
