+++ src/store/guitars/api-actions.ts
@@ -1,4 +1,4 @@
-import { loadFilteredGuitars, loadGuitars, loadGuitarsByName } from './action';
+import { loadFilteredGuitars, loadGuitars, loadGuitarsByName } from './slice';
 import { ApiRoute } from '../../constants/api-route';
 import { RootState } from '../store';
 import { ThunkAction } from '@reduxjs/toolkit';
@@ -6,6 +6,7 @@ import { Action } from 'redux';
 import { AxiosInstance } from 'axios';
 import { QueryParametersType } from '../../types/query-params';
 import { QueryParam } from '../../constants/query-param';
+import { loadTotalPageCount } from '../pagination/slice';
 
 const GUITARS_PER_PAGE = 9;
 const TotalCountHeader = 'x-total-count';
@@ -31,7 +32,8 @@ export const fetchFilteredGuitarsList = (params?: QueryParametersType): ThunkAct
   });
 
   const allPages = Math.ceil(response.headers[TotalCountHeader] / GUITARS_PER_PAGE);
-  dispatch(loadFilteredGuitars(response.data, allPages));
+  dispatch(loadTotalPageCount(allPages));
+  dispatch(loadFilteredGuitars(response.data));
 };
 
 export const fetchGuitarsListByName = (name: string): ThunkAction<Promise<void>, RootState, AxiosInstance, Action> => async (dispatch, _getState, api) => {
